% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{link_func}
\alias{link_func}
\title{This function computes our link function that respects the positivity}
\usage{
link_func(x, sigma_star)
}
\value{
returns results our link function
}
\description{
This function computes our link function that respects the positivity
}
\keyword{internal}
