% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{link_func_grad}
\alias{link_func_grad}
\title{This function computes the gradient of our link function that respects the positivity}
\usage{
link_func_grad(x, sigma_star)
}
\value{
returns the gradient of our link function
}
\description{
This function computes the gradient of our link function that respects the positivity
}
\keyword{internal}
