% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{link_func_vec}
\alias{link_func_vec}
\title{This function vectorizes link_func}
\usage{
link_func_vec(x, sigma_star)
}
\value{
returns a vector of link_func results
}
\description{
This function vectorizes link_func
}
\keyword{internal}
