% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{nmm_preprocessing}
\alias{nmm_preprocessing}
\title{This function prepares the user-provided data into the correct format}
\usage{
nmm_preprocessing(data, O)
}
\value{
returns a list containing the following components:
\item{R}{the indicator of the missing pattern, where R = 1 is reserved for fully observed}
\item{pattern}{a matrix recording the pattern}
}
\description{
This function prepares the user-provided data into the correct format
}
\keyword{internal}
