% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{prop_compute}
\alias{prop_compute}
\title{This function computes propensities of entries being fully observed}
\usage{
prop_compute(para, work_df, R, selection, sigma_star, pattern)
}
\value{
returns propensities of being fully observed
}
\description{
This function computes propensities of entries being fully observed
}
\keyword{internal}
