% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{PS}
\alias{PS}
\title{This function prepares necessary list of information for fitting IPW or AIPW}
\usage{
PS(data, R, pattern)
}
\value{
returns a list containing the following components:
\item{gamma_para}{the fitted values for the first stage regression}
\item{gamma_IF}{the influence functions of the fitted values for the first stage regression}
\item{sigma_star}{the positivity parameter used in the link function}
\item{prop}{the propensities of fully observed for each R = 1 subject}
}
\description{
This function prepares necessary list of information for fitting IPW or AIPW
}
\keyword{internal}
