% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{regress_fit}
\alias{regress_fit}
\title{This function returns the residuls after fiting a func regression, where func is the user-specified function}
\usage{
regress_fit(
  coefficients,
  regress_func,
  formula,
  data,
  weights,
  R,
  sigma_star,
  pattern,
  ...
)
}
\value{
returns the residuals of the regression by the user-specified function
}
\description{
This function returns the residuls after fiting a func regression, where func is the user-specified function
}
\keyword{internal}
